requirements.txt
Flask
PyMySQL
Flask-SQLAlchemy
Flask-Script
Flask-WTF
Flask-Migrate
py-bcrypt
python-slugify
Flask-Markdown
Flask-Uploads
openpyxl